+++ src/System/Taffybar/Widget/MPRIS2.hs
@@ -95,6 +95,23 @@ data MPRIS2PlayerWidget = MPRIS2PlayerWidget
     playerWidget :: Gtk.Widget
   }
 
+data MPRIS2PlayerControlsWidget = MPRIS2PlayerControlsWidget
+  { controlsPlayerLabel :: Gtk.Label,
+    controlsPlayerWidget :: Gtk.Widget,
+    controlsNowPlayingVar :: MV.MVar NowPlaying,
+    controlsPreviousButton :: Gtk.Button,
+    controlsPlayPauseButton :: Gtk.Button,
+    controlsPlayPauseButtonLabel :: Gtk.Label,
+    controlsNextButton :: Gtk.Button
+  }
+
+defaultMPRIS2ControlsConfig :: MPRIS2Config MPRIS2PlayerControlsWidget
+defaultMPRIS2ControlsConfig =
+  MPRIS2Config
+    { mprisWidgetWrapper = return,
+      updatePlayerWidget = simplePlayerWidgetWithControls def
+    }
+
 data SimpleMPRIS2PlayerConfig = SimpleMPRIS2PlayerConfig
   { setNowPlayingLabel :: NowPlaying -> IO T.Text,
     showPlayerWidgetFn :: NowPlaying -> IO Bool
@@ -142,6 +159,68 @@ loadIconAtSize client busName size =
                 "Failed to get image"
                 (getImageForDesktopEntry size)
           )
+
+backIconText :: T.Text
+backIconText = "⏮"
+
+playIconText :: T.Text
+playIconText = "▶"
+
+pauseIconText :: T.Text
+pauseIconText = "⏸"
+
+nextIconText :: T.Text
+nextIconText = "⏭"
+
+toggleIconText :: T.Text
+toggleIconText = "⏯"
+
+isPlaying :: NowPlaying -> Bool
+isPlaying NowPlaying {npStatus = status} = status == "Playing"
+
+canTogglePlayback :: NowPlaying -> Bool
+canTogglePlayback nowPlaying = npCanPause nowPlaying || npCanPlay nowPlaying
+
+playPauseIconText :: NowPlaying -> T.Text
+playPauseIconText nowPlaying
+  | isPlaying nowPlaying && npCanPause nowPlaying = pauseIconText
+  | npCanPlay nowPlaying = playIconText
+  | otherwise = toggleIconText
+
+runPlayPauseAction :: Client -> NowPlaying -> IO ()
+runPlayPauseAction client nowPlaying
+  | isPlaying nowPlaying && npCanPause nowPlaying =
+      void $ MPRIS2DBus.pause client (npBusName nowPlaying)
+  | not (isPlaying nowPlaying) && npCanPlay nowPlaying =
+      void $ MPRIS2DBus.play client (npBusName nowPlaying)
+  | canTogglePlayback nowPlaying =
+      void $ MPRIS2DBus.playPause client (npBusName nowPlaying)
+  | otherwise = return ()
+
+updateControlButtons :: MPRIS2PlayerControlsWidget -> NowPlaying -> IO ()
+updateControlButtons
+  MPRIS2PlayerControlsWidget
+    { controlsPreviousButton = previousButton,
+      controlsPlayPauseButton = playPauseButton,
+      controlsPlayPauseButtonLabel = playPauseButtonLabel,
+      controlsNextButton = nextButton
+    }
+  nowPlaying = do
+    Gtk.widgetSetVisible previousButton (npCanGoPrevious nowPlaying)
+    Gtk.widgetSetSensitive previousButton (npCanGoPrevious nowPlaying)
+    Gtk.widgetSetVisible playPauseButton (canTogglePlayback nowPlaying)
+    Gtk.widgetSetSensitive playPauseButton (canTogglePlayback nowPlaying)
+    Gtk.labelSetText playPauseButtonLabel (playPauseIconText nowPlaying)
+    Gtk.widgetSetVisible nextButton (npCanGoNext nowPlaying)
+    Gtk.widgetSetSensitive nextButton (npCanGoNext nowPlaying)
+
+newControlButton :: T.Text -> IO (Gtk.Button, Gtk.Label)
+newControlButton iconText = do
+  button <- Gtk.buttonNew
+  label <- Gtk.labelNew $ Just iconText
+  Gtk.containerAdd button label
+  Gtk.widgetShowAll button
+  return (button, label)
 
 -- | This is the default player widget constructor that is used to build mpris
 -- widgets. It provides only an icon and NowPlaying text.
@@ -208,9 +287,122 @@ simplePlayerWidget _ _ _ _ =
     ("nowplaying" :: String)
     >> return undefined
 
+-- | This player widget constructor extends the default MPRIS2 row with previous,
+-- play/pause, and next buttons.
+simplePlayerWidgetWithControls ::
+  SimpleMPRIS2PlayerConfig -> UpdateMPRIS2PlayerWidget MPRIS2PlayerControlsWidget
+simplePlayerWidgetWithControls
+  _
+  _
+  (Just p@MPRIS2PlayerControlsWidget {controlsPlayerWidget = widget})
+  Nothing =
+    lift $ Gtk.widgetHide widget >> return p
+simplePlayerWidgetWithControls
+  c
+  addToParent
+  Nothing
+  np@(Just nowPlaying@NowPlaying {npBusName = busName}) = do
+    ctx <- ask
+    client <- asks sessionDBusClient
+    lift $ do
+      mprisLog DEBUG "Building widget for %s" busName
+      image <- autoSizeImageNew (loadIconAtSize client busName) Gtk.OrientationHorizontal
+      playerBox <- Gtk.boxNew Gtk.OrientationHorizontal 0
+      clickArea <- Gtk.boxNew Gtk.OrientationHorizontal 0
+      controlsBox <- Gtk.boxNew Gtk.OrientationHorizontal 0
+      label <- Gtk.labelNew Nothing
+      nowPlayingVar <- MV.newMVar nowPlaying
+      (previousButton, _) <- newControlButton backIconText
+      (playPauseButton, playPauseButtonLabel) <- newControlButton toggleIconText
+      (nextButton, _) <- newControlButton nextIconText
+      _ <- widgetSetClassGI controlsBox "mpris-controls"
+      _ <- widgetSetClassGI previousButton "mpris-control"
+      _ <- widgetSetClassGI previousButton "mpris-control-previous"
+      _ <- widgetSetClassGI playPauseButton "mpris-control"
+      _ <- widgetSetClassGI playPauseButton "mpris-control-play-pause"
+      _ <- widgetSetClassGI nextButton "mpris-control"
+      _ <- widgetSetClassGI nextButton "mpris-control-next"
+      _ <- Gtk.onButtonClicked previousButton $ do
+        currentState <- MV.readMVar nowPlayingVar
+        when
+          (npCanGoPrevious currentState)
+          (void $ MPRIS2DBus.previous client (npBusName currentState))
+      _ <- Gtk.onButtonClicked playPauseButton $ do
+        currentState <- MV.readMVar nowPlayingVar
+        runPlayPauseAction client currentState
+      _ <- Gtk.onButtonClicked nextButton $ do
+        currentState <- MV.readMVar nowPlayingVar
+        when
+          (npCanGoNext currentState)
+          (void $ MPRIS2DBus.next client (npBusName currentState))
+      ebox <- Gtk.eventBoxNew
+      _ <-
+        Gtk.onWidgetButtonPressEvent ebox $
+          const $ do
+            currentState <- MV.readMVar nowPlayingVar
+            runPlayPauseAction client currentState
+            return True
+      Gtk.boxPackStart clickArea image False False 0
+      Gtk.boxPackStart clickArea label True True 0
+      Gtk.containerAdd ebox clickArea
+      Gtk.boxPackStart controlsBox previousButton False False 0
+      Gtk.boxPackStart controlsBox playPauseButton False False 0
+      Gtk.boxPackStart controlsBox nextButton False False 0
+      Gtk.boxPackStart playerBox ebox True True 0
+      Gtk.boxPackStart playerBox controlsBox False False 0
+      vFillCenter playerBox
+      addToParent playerBox
+      Gtk.widgetSetVexpand playerBox True
+      Gtk.widgetSetName playerBox $ T.pack $ formatBusName busName
+      Gtk.widgetShowAll playerBox
+      Gtk.widgetHide playerBox
+      widget <- Gtk.toWidget playerBox
+      let widgetData =
+            MPRIS2PlayerControlsWidget
+              { controlsPlayerLabel = label,
+                controlsPlayerWidget = widget,
+                controlsNowPlayingVar = nowPlayingVar,
+                controlsPreviousButton = previousButton,
+                controlsPlayPauseButton = playPauseButton,
+                controlsPlayPauseButtonLabel = playPauseButtonLabel,
+                controlsNextButton = nextButton
+              }
+      flip runReaderT ctx $
+        simplePlayerWidgetWithControls c addToParent (Just widgetData) np
+simplePlayerWidgetWithControls
+  config
+  _
+  ( Just
+      w@MPRIS2PlayerControlsWidget
+        { controlsPlayerLabel = label,
+          controlsPlayerWidget = widget,
+          controlsNowPlayingVar = nowPlayingVar
+        }
+    )
+  (Just nowPlaying) = lift $ do
+    mprisLog DEBUG "Setting state %s" nowPlaying
+    void $ MV.swapMVar nowPlayingVar nowPlaying
+    Gtk.labelSetMarkup label =<< setNowPlayingLabel config nowPlaying
+    shouldShow <- showPlayerWidgetFn config nowPlaying
+    if shouldShow
+      then Gtk.widgetShowAll widget >> updateControlButtons w nowPlaying
+      else Gtk.widgetHide widget
+    return w
+simplePlayerWidgetWithControls _ _ _ _ =
+  mprisLog
+    WARNING
+    "widget update called with no widget or %s"
+    ("nowplaying" :: String)
+    >> return undefined
+
 -- | Construct a new MPRIS2 widget using the `simplePlayerWidget` constructor.
 mpris2New :: TaffyIO Gtk.Widget
 mpris2New = mpris2NewWithConfig defaultMPRIS2Config
+
+-- | Construct a new MPRIS2 widget with transport control buttons
+-- (previous/play-pause/next) when the player advertises support for them.
+mpris2NewWithControls :: TaffyIO Gtk.Widget
+mpris2NewWithControls = mpris2NewWithConfig defaultMPRIS2ControlsConfig
 
 -- | Construct a new MPRIS2 widget with the provided configuration.
 mpris2NewWithConfig :: MPRIS2Config a -> TaffyIO Gtk.Widget